+++ miracle/mysite/config/urls.py
@@ -23,5 +23,5 @@ urlpatterns = [
     path('rutines/', include('rutineapp.urls')),
     path('accounts/', include('accountapp.urls')),
     path('profiles/', include('profileapp.urls')),
-    path('calendar/', include('calendarapp.urls')),
+    path('calendar/', include('newcalendarapp.urls')),
 ]
