+++ miracle/mysite/newcalendarapp/forms.py
@@ -6,11 +6,9 @@ class EventForm(ModelForm):
         model = Event
         widgets = {
             'start_time': DateInput(attrs={'type': 'date'}, format='%Y-%m-%d'),
-            'end_time': DateInput(attrs={'type': 'date'}, format='%Y-%m-%d'),
         }
         fields = '__all__'
     
     def __init__(self, *args, **kwargs):
         super(EventForm, self).__init__(*args, **kwargs)
         self.fields['start_time'].input_formats = ('%Y-%m-%d',)
-        self.fields['end_time'].input_formats = ('%Y-%m-%d',)
