+++ miracle/mysite/newcalendarapp/tests.py
