+++ miracle/mysite/newcalendarapp/views.py
