+++ miracle/mysite/rutineapp/forms.py
@@ -3,8 +3,8 @@ from django import forms
 from rutineapp.models import Rutine
 
 class RutineCreationForm(ModelForm):
-    content = forms.CharField(widget=forms.Textarea(attrs={'class':'editable', 'style':'height:auto;'}))
+    # content = forms.CharField(widget=forms.Textarea(attrs={'class':'editable', 'style':'height:auto;'}))
 
     class Meta:
         model = Rutine
-        fields = ['title','content']
+        fields = ['title']
