+++ miracle/mysite/rutineapp/models.py
@@ -3,4 +3,8 @@ from django.db import models
 # Create your models here.
 class Rutine(models.Model):
     title = models.CharField(max_length=200)
-    content = models.TextField(null=True)
+    color = models.CharField(max_length=150,default='red',null=True)
+
+
+    def __str__(self):
+        return f'{self.title}, {self.color}'
